backend/requirements.txt
fastapi[all]==0.104.1
sqlmodel==0.0.14
psycopg2-binary==2.9.9
uvicorn==0.24.0
python-dotenv==1.0.0